requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==2.3.0
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.33.0
django-rest-swagger==2.1.2
djangorestframework==3.6.3
dj-database-url==0.4.2
dj-static==0.0.6
gunicorn==19.7.0
idna==2.0
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2017.2
python3-openid==3.1.0
packaging==16.8
psycopg2>=2.6.1
pyparsing==2.2.0
requests==2.9.1
simplejson==3.11.1
six==1.10.0
static3==0.7.0
uritemplate==3.0.0
uritemplate.py==0.3.0
urllib3==1.22